Create Matrix Operations on 2d Lists
In this assignment, you will implement several basic linear algebra operations. 
The underlying storage of data elements will be as a 2-dimensional list within Python. Each operation will
be implemented as its own a function, and you will be responsible for running unit tests on each function 
to ensure it works correctly. As always, you must write a concise and descriptive doc string for each function, 
and include comments as necessary throughout the code.

NOTE: I did not perform well on this homework. Hence the code written here may be incorrect and will be changed/
updated over the coming days.

Thank you Professor Aaron Stevens for granting the authorization to post privately on Github.