Create User Defined Class- Bonds
Thank you professor Aaron Stevens for granting me the permission to post classwork on Github privately.

In this assignment, you will develop a class to model a bond investment instrument. The bond has a maturity
value (in dollars), maturity time (in years), coupon rate, coupon (compounding) frequency, current price, and
yield to maturity (YTM). To keep price and YTM in synch, any changes in price will force the YTM to recalculate, 
and changes to YTM will force the price to recalculate.